requirements.txt
kafka-python==2.0.2
pandas==2.1.1
SQLAlchemy==2.1.0
psycopg2-binary==2.9.8
PyYAML==6.0